[feat] Moved the table at the home page to its own component and made entire rows clickable (#89)
