+++ frontend/src/pages/CVDatabase.tsx
@@ -1,7 +1,7 @@
 import { useFetchCandidates } from "@/hooks/useFetchCandidates";
 import { useState } from "react";
 import PdfPreviewOverlay from "../components/PdfPreviewOverlay";
-import { AddNewCVModal } from "@/components/addNewCv/AddNewCVModal";
+import { AddNewCVModal } from "@/components/AddNewCv/AddNewCVModal";
 import { Spinner } from "@/components/ui/spinner";
 import type { CandidatePreview } from "@/types/candidate";
 import ErrorBox from "@/components/ErrorBox";
